Threading Support Added
Made whole module thread safe, added option for Trader.trade() to run
the method in a seperate thread which allows the users program to make
calculations based on the current price or the users input ect. This
trading thread is set up to be killed with ctrl+c with the daemon line.
Also cleaned up the code in a few places. Examples are shown but
commented out in the testTrader.py file. Open to any suggestions or
improvements.