One Starts Another Trade
Allows you to create a sequence of trades that will only execute once
the prior order is executed. I am slightly concerned that if you buy
coins then try to sell them in the next order that the coins won't be in
your wallet before the sell order which would cause a CoinbaseError
although you should have time assuming you are using a limit order.